day24_Chroma_Fullstack/requirements.txt
fastapi
uvicorn
streamlit
chromadb
torch
transformers
pillow
python-multipart